feat(report): unify ReportForm, guarantee event page refresh after submission\n\n- Extracted a shared ReportForm component for both modal and in-page use\n- Ensured consistent submission, reset, and redirect/refresh logic\n- Now uses router.reload() to guarantee the event page always refreshes and shows new reports\n- Fixes #84\n\nThis closes #84 by ensuring users are always redirected to a refreshed event page after submitting a report, regardless of how the form is accessed (modal or in-page).
